updated campaignId to type Long

+++ src/main/java/sg/edu/smu/cs301/group3/campaignms/beans/NotificationBean.java
@@ -15,7 +15,7 @@ public class NotificationBean {
     private String notificationMessage;
 
     @JsonProperty("campaign_id")
-    private int campaignId;
+    private Long campaignId;
 
     public static NotificationBean fromNotificationModel(Notification notification) {
         return NotificationBean.builder()
