+++ src/main/java/sg/edu/smu/cs301/group3/campaignms/repository/NotificationsRepository.java
@@ -6,6 +6,6 @@ import sg.edu.smu.cs301.group3.campaignms.model.Notification;
 import java.util.List;
 
 public interface NotificationsRepository extends JpaRepository<Notification, Long> {
-    List<Notification> getNotificationsByCampaignId(int campaignId);
-    void deleteAllByCampaignId(int campaignId);
+    List<Notification> getNotificationsByCampaignId(Long campaignId);
+    void deleteAllByCampaignId(Long campaignId);
 }
